+++ public/index.html
@@ -4,7 +4,7 @@
   <head>
     <meta charset="utf-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1" />
-    <title>중고마켓 페이지</title>
+    <title>판다마켓</title>
     <link rel="icon" href="./favicon.ico" />
     <link
       rel="stylesheet"
