+++ src/Main.js
@@ -1,13 +1,15 @@
 import { BrowserRouter, Routes, Route } from "react-router-dom";
 import App from './components/App';
+import HomePage from "./pages/HomePage";
 
 
 function Main() {
   return (
     <BrowserRouter>
-        <Routes>
-          <Route path="items" element={<App />}></Route>
-        </Routes>
+      <Routes>
+        <Route path="/" element={<HomePage />}></Route>
+        <Route path="items" element={<App />}></Route>
+      </Routes>  
     </BrowserRouter>
   );
