+++ src/components/Pagination.css
@@ -1,0 +1,29 @@
+.Pagination {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  gap: 4px;
+}
+.pagination-prev,
+.pagination-num,
+.pagination-next {
+  padding: 7px 16px;
+  border: 1px solid #E5E7EB;
+  border-radius: 50px;
+  font-weight: 600;
+  font-size: 16px;
+  line-height: 26px;
+  color: #6B7280;
+  cursor: pointer;
+}
+.pagination-prev:hover,
+.pagination-num:hover,
+.pagination-next:hover {
+  background-color: #2F80ED;
+  border: none;
+  color: #F9FAFB;
+}
+.pagination-num.active {
+  background-color: #007bff;
+  color: white;
+}
