+++ src/components/ProductList.css
@@ -100,42 +100,6 @@
 
 
 
-
-.Pagination {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  gap: 4px;
-}
-.pagination-prev,
-.pagination-num,
-.pagination-next {
-  padding: 7px 16px;
-  border: 1px solid #E5E7EB;
-  border-radius: 50px;
-  font-weight: 600;
-  font-size: 16px;
-  line-height: 26px;
-  color: #6B7280;
-  cursor: pointer;
-}
-.pagination-prev:hover,
-.pagination-num:hover,
-.pagination-next:hover {
-  background-color: #2F80ED;
-  border: none;
-  color: #F9FAFB;
-}
-.pagination-num.active {
-  background-color: #007bff;
-  color: white;
-}
-
-
-
-
-
-
 @media screen and (min-width: 787px) and (max-width: 1460px) {
   .ProductList {
     padding: 24px 24px 72px 24px;
