+++ src/pages/HomePage.jsx
@@ -1,0 +1,12 @@
+import Nav from "../components/Nav";
+
+function HomePage() {
+  return (
+    <div className="HomePage">
+      <Nav/>
+      <div><h1>HOME</h1></div>
+    </div>
+  );
+}
+
+export default HomePage;
